chore: move all theme configs in one folder
